+++ app/services/paper_factory.rb
@@ -59,7 +59,6 @@ class PaperFactory
       creator: creator,
       title: task_template.title,
       body: task_template.template,
-      old_role: journal_task_type.old_role,
       notify: false
     )
     task_class.task_added_to_paper(paper)