+++ engines/tahi_standard_tasks/app/models/tahi_standard_tasks/publishing_related_questions_task.rb
@@ -2,7 +2,7 @@ module TahiStandardTasks
   class PublishingRelatedQuestionsTask < Task
     include MetadataTask
 
-    DEFAULT_TITLE = 'Additional Information'
-    DEFAULT_ROLE = 'author'
+    DEFAULT_TITLE = 'Additional Information'.freeze
+    DEFAULT_ROLE_HINT = 'author'.freeze
   end
 end