+++ engines/tahi_standard_tasks/app/models/tahi_standard_tasks/reviewer_report_task.rb
@@ -4,18 +4,17 @@ module TahiStandardTasks
   # up in the "ALL REVIEWS COMPLETE" query, it should inherit from
   # ReviewerReportTask.
   class ReviewerReportTask < Task
-    DEFAULT_TITLE = 'Reviewer Report'
-    DEFAULT_ROLE = 'reviewer'
+    DEFAULT_TITLE = 'Reviewer Report'.freeze
+    DEFAULT_ROLE_HINT = 'reviewer'.freeze
     SYSTEM_GENERATED = true
 
     before_create :assign_to_draft_decision
     has_many :decisions, -> { uniq }, through: :paper
 
-    # Overrides Task#restore_defaults to be only restore +old_role+. This
+    # Overrides Task#restore_defaults to not restore +title+. This
     # will never update +title+ as that is dynamically determined. If you
     # need to change the reviewer report title write a data migration.
     def self.restore_defaults
-      update_all(old_role: self::DEFAULT_ROLE)
     end
 
     # find_or_build_answer_for(...) will return the associated answer for this
@@ -76,9 +75,7 @@ module TahiStandardTasks
       previous_decision_ids = body["previous_decision_ids"] || []
       current_decision_id = body["decision_id"]
 
-      if current_decision_id
-        previous_decision_ids.push current_decision_id
-      end
+      previous_decision_ids.push current_decision_id if current_decision_id
 
       update_body(
         "decision_id" => new_decision.try(:id),