+++ engines/tahi_standard_tasks/spec/factories/register_decision_task_factory.rb
@@ -3,6 +3,5 @@ FactoryGirl.define do
     paper
     phase
     title "Register Decision"
-    old_role "editor"
   end
 end