+++ engines/tahi_standard_tasks/spec/models/financial_disclosure_task_spec.rb
@@ -5,7 +5,6 @@ describe TahiStandardTasks::FinancialDisclosureTask do
 
   describe '.restore_defaults' do
     it_behaves_like '<Task class>.restore_defaults update title to the default'
-    it_behaves_like '<Task class>.restore_defaults update old_role to the default'
   end
 
   describe '#funders association' do